Split render/status.rs into per-concern modules
