--- src/render/status.rs
@@ -1,1644 +1,0 @@
-//! Status bar and input popup rendering
-
-use std::cell::RefCell;
-use std::path::{Path, PathBuf};
-use std::sync::OnceLock;
-use std::time::SystemTime;
-
-use ratatui::{
-    layout::{Constraint, Direction, Layout, Rect},
-    style::{Color, Modifier, Style},
-    text::{Line, Span},
-    widgets::{Block, Borders, Clear, Paragraph},
-    Frame,
-};
-
-use std::fs::File;
-use std::io::{BufRead, BufReader};
-
-use super::layout::LayoutEngine;
-use super::theme::theme;
-use crate::core::{
-    AppState, InputPurpose, PendingAction, PreviewDisplayMode, SortMode, UiDensity, ViewMode,
-};
-use crate::integrate::{humanize_tokens, BudgetSeverity};
-
-/// Build the spans rendered for the context budget bar.
-///
-/// Returns an empty vector when nothing is selected.
-fn budget_segment_spans(state: &AppState, density: UiDensity) -> Vec<Span<'static>> {
-    if state.selected_paths.is_empty() {
-        return Vec::new();
-    }
-    let t = theme();
-    let used = state.known_token_total();
-    let pending = state.pending_token_count();
-    let window = state.budget_model.window_tokens();
-    let percent = if window > 0 {
-        ((used as f64 / window as f64) * 100.0).round() as u32
-    } else {
-        0
-    };
-    let severity = BudgetSeverity::from_usage(used, window);
-    let color = match severity {
-        BudgetSeverity::Ok => t.git_staged,
-        BudgetSeverity::Warn => t.warning,
-        BudgetSeverity::Hot => t.git_conflict,
-        BudgetSeverity::Over => Color::Magenta,
-    };
-    let used_s = humanize_tokens(used);
-    let win_s = humanize_tokens(window);
-    let pending_marker = if pending > 0 {
-        format!("+{}?", pending)
-    } else {
-        String::new()
-    };
-    let text = match density {
-        UiDensity::Full => format!(
-            "[ctx {}f {}{}/{} {} {}%]",
-            state.selected_paths.len(),
-            used_s,
-            if pending_marker.is_empty() {
-                String::new()
-            } else {
-                format!(" {}", pending_marker)
-            },
-            win_s,
-            state.budget_model.short_label(),
-            percent,
-        ),
-        UiDensity::Compact => format!(
-            "[ctx {}{}/{} {}%]",
-            used_s,
-            if pending_marker.is_empty() {
-                String::new()
-            } else {
-                format!(" {}", pending_marker)
-            },
-            win_s,
-            percent,
-        ),
-        UiDensity::Narrow => format!("[ctx {}/{}]", used_s, win_s),
-        UiDensity::Ultra => format!("[{}]", used_s),
-    };
-    vec![Span::styled(text, Style::default().fg(color))]
-}
-
-/// Render the status bar with adaptive layout based on screen width
-/// Compose the status-bar message, giving AI activity precedence when a
-/// recent event is available. Returns an owned `String` because we may need
-/// to concatenate with the regular `state.message`.
-pub fn effective_status_message(state: &AppState, fallback: &str) -> String {
-    let indicator = super::ai_activity::status_indicator(state);
-    match (indicator, state.message.as_deref()) {
-        (Some(ind), Some(msg)) => format!("{} · {}", ind, msg),
-        (Some(ind), None) => ind,
-        (None, Some(msg)) => msg.to_string(),
-        (None, None) => fallback.to_string(),
-    }
-}
-
-pub fn render_status_bar(
-    frame: &mut Frame,
-    state: &AppState,
-    focused_path: Option<&PathBuf>,
-    area: Rect,
-) {
-    // Check if peek mode is enabled - render peek preview instead of normal status
-    if state.preview_display_mode == PreviewDisplayMode::Peek {
-        render_peek_status(frame, state, focused_path, area);
-        return;
-    }
-
-    let density = state.ui_density_for_width(area.width);
-    let layout = LayoutEngine::from_rect_with_density(area, density);
-
-    // Adaptive layout based on UI density
-    match layout.density {
-        UiDensity::Ultra => render_ultra_compact_status(frame, state, area),
-        UiDensity::Narrow => render_compact_status(frame, state, focused_path, area),
-        UiDensity::Compact => render_narrow_status(frame, state, focused_path, area),
-        UiDensity::Full => render_full_status(frame, state, focused_path, area),
-    }
-}
-
-/// Render peek mode status bar (shows file preview in status area)
-fn render_peek_status(
-    frame: &mut Frame,
-    state: &AppState,
-    focused_path: Option<&PathBuf>,
-    area: Rect,
-) {
-    let t = theme();
-    let density = state.ui_density_for_width(area.width);
-    let layout = LayoutEngine::from_rect_with_density(area, density);
-    let max_preview_lines = layout.peek_preview_lines();
-
-    // Build header line with minimal info based on density
-    let mut header_spans = Vec::new();
-
-    // Peek indicator
-    header_spans.push(Span::styled(
-        "P",
-        Style::default().fg(t.info).add_modifier(Modifier::BOLD),
-    ));
-
-    // Selection count
-    if !state.selected_paths.is_empty() {
-        header_spans.push(Span::raw(" "));
-        header_spans.push(Span::styled(
-            format!("{}*", state.selected_paths.len()),
-            Style::default().fg(t.mark),
-        ));
-    }
-
-    // Git branch (abbreviated based on density)
-    if let Some(branch) = state.git_status.as_ref().and_then(|g| g.branch()) {
-        let max_branch_len = match layout.density {
-            UiDensity::Ultra => 4,
-            UiDensity::Narrow => 6,
-            _ => 8,
-        };
-        header_spans.push(Span::raw(" "));
-        let branch_abbrev = if branch.len() > max_branch_len {
-            format!("\u{e0a0}{}…", &branch[..max_branch_len - 1])
-        } else {
-            format!("\u{e0a0}{}", branch)
-        };
-        header_spans.push(Span::styled(
-            branch_abbrev,
-            Style::default().fg(t.git_staged),
-        ));
-    }
-
-    // File name (only if enough space)
-    if let Some(path) = focused_path {
-        if let Some(name) = path.file_name().and_then(|n| n.to_str()) {
-            let separator = match layout.density {
-                UiDensity::Ultra => " ",
-                _ => " │ ",
-            };
-            let max_name_len = match layout.density {
-                UiDensity::Ultra => (area.width as usize).saturating_sub(12),
-                UiDensity::Narrow => (area.width as usize).saturating_sub(15),
-                _ => (area.width as usize).saturating_sub(20),
-            };
-            if max_name_len > 3 {
-                header_spans.push(Span::styled(separator, Style::default().fg(t.git_ignored)));
-                let display_name = if name.len() > max_name_len {
-                    format!("{}…", &name[..max_name_len.saturating_sub(1)])
-                } else {
-                    name.to_string()
-                };
-                header_spans.push(Span::styled(
-                    display_name,
-                    Style::default().fg(t.border_active),
-                ));
-            }
-        }
-    }
-
-    // Read first few lines of the file for peek preview
-    let preview_lines = if let Some(path) = focused_path {
-        read_peek_lines(
-            path,
-            max_preview_lines.min(area.height.saturating_sub(2) as usize),
-        )
-    } else {
-        vec!["(No file)".to_string()]
-    };
-
-    let mut content = vec![Line::from(header_spans)];
-    for line in preview_lines {
-        // Truncate long lines
-        let max_width = area.width.saturating_sub(2) as usize;
-        let display = if line.len() > max_width {
-            format!("{}…", &line[..max_width.saturating_sub(1)])
-        } else {
-            line
-        };
-        content.push(Line::from(Span::styled(
-            display,
-            Style::default().fg(t.git_ignored),
-        )));
-    }
-
-    // Compact title for ultra mode
-    let title = match layout.density {
-        UiDensity::Ultra => " P ",
-        _ => " Peek ",
-    };
-
-    let widget = Paragraph::new(content).block(
-        Block::default()
-            .borders(Borders::ALL)
-            .border_style(Style::default().fg(t.border_active))
-            .title(title),
-    );
-    frame.render_widget(widget, area);
-}
-
-/// Read first few lines of a file for peek preview
-fn read_peek_lines(path: &PathBuf, max_lines: usize) -> Vec<String> {
-    if path.is_dir() {
-        return vec!["(directory)".to_string()];
-    }
-
-    let file = match File::open(path) {
-        Ok(f) => f,
-        Err(_) => return vec!["(cannot read)".to_string()],
-    };
-
-    let reader = BufReader::new(file);
-    let mut lines = Vec::new();
-
-    for line in reader.lines().take(max_lines) {
-        match line {
-            Ok(l) => {
-                // Skip empty lines at the start
-                if lines.is_empty() && l.trim().is_empty() {
-                    continue;
-                }
-                // Replace tabs with spaces
-                lines.push(l.replace('\t', "  "));
-            }
-            Err(_) => {
-                if lines.is_empty() {
-                    return vec!["(binary file)".to_string()];
-                }
-                break;
-            }
-        }
-    }
-
-    if lines.is_empty() {
-        vec!["(empty file)".to_string()]
-    } else {
-        lines
-    }
-}
-
-/// Render ultra-compact status bar for extremely narrow screens (< 25 chars)
-/// Shows minimal info: `? 3* ⎇m` (help, selection count, git branch)
-/// Optimized for 20-24 character width terminals
-fn render_ultra_compact_status(frame: &mut Frame, state: &AppState, area: Rect) {
-    let t = theme();
-    let mut spans = Vec::new();
-    let inner_width = area.width.saturating_sub(2) as usize; // Account for borders
-
-    // For ultra-narrow, prioritize: selection > branch > sort > filter > message
-    let selected_count = state.selected_paths.len();
-
-    // Selection count first (most important in ultra mode)
-    if selected_count > 0 {
-        spans.push(Span::styled(
-            format!("{}*", selected_count),
-            Style::default().fg(t.mark),
-        ));
-    }
-
-    // Git branch (very abbreviated)
-    if let Some(branch) = state.git_status.as_ref().and_then(|g| g.branch()) {
-        if !spans.is_empty() {
-            spans.push(Span::raw(" "));
-        }
-        // Max 4 chars for branch in ultra mode
-        let branch_abbrev = if branch.len() > 4 {
-            format!("\u{e0a0}{}…", &branch[..3])
-        } else {
-            format!("\u{e0a0}{}", branch)
-        };
-        spans.push(Span::styled(
-            branch_abbrev,
-            Style::default().fg(t.git_staged),
-        ));
-    }
-
-    // Sort mode indicator (single char, only if not default)
-    if state.sort_mode != SortMode::Name {
-        if !spans.is_empty() {
-            spans.push(Span::raw(" "));
-        }
-        spans.push(Span::styled(
-            state.sort_mode.short_name(),
-            Style::default().fg(t.git_conflict),
-        ));
-    }
-
-    // Filter indicator (just an icon)
-    if state.filter_pattern.is_some() {
-        if !spans.is_empty() {
-            spans.push(Span::raw(" "));
-        }
-        spans.push(Span::styled("\u{f0b0}", Style::default().fg(t.warning)));
-    }
-
-    // Budget segment (single bracketed token count)
-    let budget_spans = budget_segment_spans(state, UiDensity::Ultra);
-    if !budget_spans.is_empty() {
-        if !spans.is_empty() {
-            spans.push(Span::raw(" "));
-        }
-        spans.extend(budget_spans);
-    }
-
-    // Help hint only if there's space
-    let current_width: usize = spans.iter().map(|s| s.width()).sum();
-    if current_width < inner_width.saturating_sub(2) && spans.is_empty() {
-        spans.push(Span::styled("?", Style::default().fg(t.info)));
-    }
-
-    // Message (only if there's significant space left). AI activity indicator
-    // takes precedence when a recent event is available.
-    let message = effective_status_message(state, "");
-    if !message.is_empty() {
-        let used_width: usize = spans.iter().map(|s| s.width()).sum();
-        let available = inner_width.saturating_sub(used_width + 1);
-        if available > 3 {
-            spans.push(Span::raw(" "));
-            let truncated = if message.len() > available {
-                format!("{}…", &message[..available.saturating_sub(1)])
-            } else {
-                message.clone()
-            };
-            spans.push(Span::raw(truncated));
-        }
-    }
-
-    let content = Line::from(spans);
-    let widget = Paragraph::new(content).block(Block::default().borders(Borders::ALL));
-    frame.render_widget(widget, area);
-}
-
-/// Render compact status bar for very narrow screens (< 60 chars)
-/// Shows only the most essential information in a single panel
-fn render_compact_status(
-    frame: &mut Frame,
-    state: &AppState,
-    focused_path: Option<&PathBuf>,
-    area: Rect,
-) {
-    let t = theme();
-
-    // Build compact content: "? | 1.2KB | main | Sel:3"
-    let mut spans = Vec::new();
-
-    // Help or message (highest priority)
-    let message_owned = effective_status_message(state, "?");
-    let message = message_owned.as_str();
-    spans.push(Span::raw(format!(" {}", message)));
-
-    // File size only (no modification time)
-    if let Some(size) = focused_path.and_then(|p| get_file_size_only(p.as_path())) {
-        spans.push(Span::styled(" | ", Style::default().fg(t.git_ignored)));
-        spans.push(Span::raw(size));
-    }
-
-    // Git branch (abbreviated, medium priority)
-    if let Some(branch) = state.git_status.as_ref().and_then(|g| g.branch()) {
-        spans.push(Span::styled(" | ", Style::default().fg(t.git_ignored)));
-        spans.push(Span::styled(
-            format!("\u{e0a0}{}", branch),
-            Style::default().fg(t.git_staged),
-        ));
-    }
-
-    // Selection count (abbreviated)
-    let selected_count = state.selected_paths.len();
-    if selected_count > 0 {
-        spans.push(Span::styled(" | ", Style::default().fg(t.git_ignored)));
-        spans.push(Span::raw(format!("Sel:{}", selected_count)));
-    }
-
-    let budget_spans = budget_segment_spans(state, UiDensity::Compact);
-    if !budget_spans.is_empty() {
-        spans.push(Span::styled(" | ", Style::default().fg(t.git_ignored)));
-        spans.extend(budget_spans);
-    }
-
-    let content = Line::from(spans);
-    let widget = Paragraph::new(content).block(Block::default().borders(Borders::ALL));
-    frame.render_widget(widget, area);
-}
-
-/// Render narrow status bar for medium screens (60-99 chars)
-/// Shows abbreviated information in two panels
-fn render_narrow_status(
-    frame: &mut Frame,
-    state: &AppState,
-    focused_path: Option<&PathBuf>,
-    area: Rect,
-) {
-    // Dynamic split: adjust based on content
-    let chunks = Layout::default()
-        .direction(Direction::Horizontal)
-        .constraints([Constraint::Percentage(45), Constraint::Percentage(55)])
-        .split(area);
-
-    let t = theme();
-
-    // Left panel: message/help + git branch
-    let mut left_spans = Vec::new();
-
-    // Watch indicator (if enabled, keep it compact)
-    if state.watch_enabled {
-        left_spans.push(Span::styled("\u{f06e} ", Style::default().fg(t.info)));
-    }
-
-    // Git branch (abbreviated)
-    if let Some(branch) = state.git_status.as_ref().and_then(|g| g.branch()) {
-        left_spans.push(Span::styled(
-            format!("\u{e0a0}{} |", branch),
-            Style::default().fg(t.git_staged),
-        ));
-    }
-
-    // Sort mode (abbreviated, only if non-default)
-    if state.sort_mode != SortMode::Name {
-        left_spans.push(Span::styled(
-            format!("\u{f0dc}{}|", state.sort_mode.short_name()),
-            Style::default().fg(t.git_conflict),
-        ));
-    }
-
-    // Search matches (abbreviated)
-    if let Some((current, total)) = state.search_matches {
-        left_spans.push(Span::styled(
-            format!("{}/{}|", current, total),
-            Style::default().fg(t.border_active),
-        ));
-    }
-
-    // Help or message
-    let message_owned = effective_status_message(state, "? help");
-    let message = message_owned.as_str();
-    left_spans.push(Span::raw(format!(" {}", message)));
-
-    let left_content = Line::from(left_spans);
-    let left_widget = Paragraph::new(left_content).block(Block::default().borders(Borders::ALL));
-    frame.render_widget(left_widget, chunks[0]);
-
-    // Right panel: file info + selection (abbreviated)
-    let file_info = focused_path
-        .map(|p| p.as_path())
-        .and_then(get_file_info_narrow)
-        .unwrap_or_else(|| "--".to_string());
-
-    let selected_count = state.selected_paths.len();
-    let clipboard_info = state
-        .clipboard
-        .as_ref()
-        .map(|c| {
-            if c.is_cut() {
-                format!(" | Cut:{}", c.paths().len())
-            } else {
-                format!(" | Cp:{}", c.paths().len())
-            }
-        })
-        .unwrap_or_default();
-
-    let stats_text = format!(
-        "{}{}{}",
-        file_info,
-        if selected_count > 0 {
-            format!(" | Sel:{}", selected_count)
-        } else {
-            String::new()
-        },
-        clipboard_info
-    );
-    let mut stats_spans: Vec<Span<'static>> = vec![Span::raw(stats_text)];
-    let budget_spans = budget_segment_spans(state, UiDensity::Narrow);
-    if !budget_spans.is_empty() {
-        stats_spans.push(Span::raw(" "));
-        stats_spans.extend(budget_spans);
-    }
-    let stats_widget =
-        Paragraph::new(Line::from(stats_spans)).block(Block::default().borders(Borders::ALL));
-    frame.render_widget(stats_widget, chunks[1]);
-}
-
-/// Render full status bar for wide screens (>= 100 chars)
-/// Original implementation with full information display
-fn render_full_status(
-    frame: &mut Frame,
-    state: &AppState,
-    focused_path: Option<&PathBuf>,
-    area: Rect,
-) {
-    let chunks = Layout::default()
-        .direction(Direction::Horizontal)
-        .constraints([Constraint::Percentage(50), Constraint::Percentage(50)])
-        .split(area);
-
-    // Left: message or help hint, with mode, git branch, watch, filter, sort, and search indicators
-    let mode_indicator = if state.select_mode {
-        if state.multi_select {
-            "\u{f05c}\u{f05c} " // Multi-select icon
-        } else {
-            "\u{f05c} " // Select icon (nf-fa-circle_o)
-        }
-    } else if state.pick_mode {
-        "\u{f046} " // Pick icon (nf-fa-check_square_o)
-    } else {
-        ""
-    };
-
-    let watch_indicator = if state.watch_enabled {
-        "\u{f06e} " // Eye icon (nf-fa-eye) for file watching
-    } else {
-        ""
-    };
-
-    let filter_indicator = state
-        .filter_pattern
-        .as_ref()
-        .map(|p| format!("\u{f0b0} {} |", p)) // Filter icon
-        .unwrap_or_default();
-
-    let branch_info = state
-        .git_status
-        .as_ref()
-        .and_then(|g| g.branch())
-        .map(|b| format!("\u{e0a0} {} |", b)) // Git branch icon
-        .unwrap_or_default();
-
-    // Sort mode indicator (only show if not default)
-    let sort_indicator = if state.sort_mode != SortMode::Name {
-        format!("\u{f0dc} {} |", state.sort_mode.display_name()) // Sort icon
-    } else {
-        String::new()
-    };
-
-    // Search match info
-    let search_indicator = state
-        .search_matches
-        .map(|(current, total)| format!("{}/{} matches |", current, total))
-        .unwrap_or_default();
-
-    let t = theme();
-    let message_owned = effective_status_message(state, "? for help");
-    let message = message_owned.as_str();
-    let left_content = Line::from(vec![
-        Span::styled(mode_indicator, Style::default().fg(t.selection)),
-        Span::styled(watch_indicator, Style::default().fg(t.info)),
-        Span::styled(filter_indicator, Style::default().fg(t.warning)),
-        Span::styled(branch_info, Style::default().fg(t.git_staged)),
-        Span::styled(sort_indicator, Style::default().fg(t.git_conflict)),
-        Span::styled(search_indicator, Style::default().fg(t.border_active)),
-        Span::raw(format!(" {}", message)),
-    ]);
-    let msg_widget = Paragraph::new(left_content).block(Block::default().borders(Borders::ALL));
-    frame.render_widget(msg_widget, chunks[0]);
-
-    // Right: file info + selection stats
-    let file_info = focused_path
-        .map(|p| p.as_path())
-        .and_then(get_file_info)
-        .unwrap_or_else(|| "--".to_string());
-
-    let selected_count = state.selected_paths.len();
-    let clipboard_info = state
-        .clipboard
-        .as_ref()
-        .map(|c| {
-            if c.is_cut() {
-                format!(" | Cut: {}", c.paths().len())
-            } else {
-                format!(" | Copied: {}", c.paths().len())
-            }
-        })
-        .unwrap_or_default();
-
-    let stats_text = format!(
-        "{}{}{}",
-        file_info,
-        if selected_count > 0 {
-            format!(" | Selected: {}", selected_count)
-        } else {
-            String::new()
-        },
-        clipboard_info
-    );
-    let mut stats_spans: Vec<Span<'static>> = vec![Span::raw(stats_text)];
-    let budget_spans = budget_segment_spans(state, UiDensity::Full);
-    if !budget_spans.is_empty() {
-        stats_spans.push(Span::raw(" "));
-        stats_spans.extend(budget_spans);
-    }
-    let stats_widget =
-        Paragraph::new(Line::from(stats_spans)).block(Block::default().borders(Borders::ALL));
-    frame.render_widget(stats_widget, chunks[1]);
-}
-
-/// Get file size and modification time as a formatted string (full display)
-/// Cached file metadata to avoid repeated stat() calls on the same path.
-struct FileInfoCache {
-    path: PathBuf,
-    size: u64,
-    is_dir: bool,
-    modified: Option<SystemTime>,
-}
-
-thread_local! {
-    static FILE_INFO_CACHE: RefCell<Option<FileInfoCache>> = const { RefCell::new(None) };
-}
-
-/// Run `f` with cached metadata for `path`. Only calls stat() when the path changes.
-fn with_cached_metadata<T>(path: &Path, f: impl FnOnce(&FileInfoCache) -> T) -> Option<T> {
-    FILE_INFO_CACHE.with(|cache| {
-        let mut cache = cache.borrow_mut();
-        let needs_update = cache.as_ref().is_none_or(|c| c.path != *path);
-        if needs_update {
-            let metadata = path.metadata().ok()?;
-            *cache = Some(FileInfoCache {
-                path: path.to_path_buf(),
-                size: metadata.len(),
-                is_dir: metadata.is_dir(),
-                modified: metadata.modified().ok(),
-            });
-        }
-        cache.as_ref().map(f)
-    })
-}
-
-fn get_file_info(path: &Path) -> Option<String> {
-    with_cached_metadata(path, |c| {
-        let size_str = if c.is_dir {
-            "--".to_string()
-        } else {
-            format_size(c.size)
-        };
-        let mtime_str = c
-            .modified
-            .map(format_relative_time)
-            .unwrap_or_else(|| "--".to_string());
-        format!("{} · {}", size_str, mtime_str)
-    })
-}
-
-/// Get file size and abbreviated modification time (narrow display)
-fn get_file_info_narrow(path: &Path) -> Option<String> {
-    with_cached_metadata(path, |c| {
-        let size_str = if c.is_dir {
-            "--".to_string()
-        } else {
-            format_size(c.size)
-        };
-        let mtime_str = c
-            .modified
-            .map(format_relative_time_short)
-            .unwrap_or_else(|| "--".to_string());
-        format!("{} · {}", size_str, mtime_str)
-    })
-}
-
-/// Get file size only (compact display)
-fn get_file_size_only(path: &Path) -> Option<String> {
-    with_cached_metadata(path, |c| {
-        if c.is_dir {
-            "--".to_string()
-        } else {
-            format_size(c.size)
-        }
-    })
-}
-
-/// Format file size in human-readable format
-fn format_size(bytes: u64) -> String {
-    const KB: u64 = 1024;
-    const MB: u64 = KB * 1024;
-    const GB: u64 = MB * 1024;
-
-    if bytes >= GB {
-        format!("{:.1} GB", bytes as f64 / GB as f64)
-    } else if bytes >= MB {
-        format!("{:.1} MB", bytes as f64 / MB as f64)
-    } else if bytes >= KB {
-        format!("{:.1} KB", bytes as f64 / KB as f64)
-    } else {
-        format!("{} B", bytes)
-    }
-}
-
-/// Format time as relative (e.g., "2h ago", "Yesterday", "Jan 30")
-fn format_relative_time(time: SystemTime) -> String {
-    let now = SystemTime::now();
-    let duration = match now.duration_since(time) {
-        Ok(d) => d,
-        Err(_) => return "Future".to_string(),
-    };
-
-    let secs = duration.as_secs();
-    let mins = secs / 60;
-    let hours = mins / 60;
-    let days = hours / 24;
-
-    if secs < 60 {
-        "Just now".to_string()
-    } else if mins < 60 {
-        format!("{}m ago", mins)
-    } else if hours < 24 {
-        format!("{}h ago", hours)
-    } else if days == 1 {
-        "Yesterday".to_string()
-    } else if days < 7 {
-        format!("{}d ago", days)
-    } else {
-        // Use date format for older files
-        use std::time::UNIX_EPOCH;
-        let timestamp = time
-            .duration_since(UNIX_EPOCH)
-            .unwrap_or_default()
-            .as_secs();
-        format_date_from_timestamp(timestamp)
-    }
-}
-
-/// Format time as short relative (e.g., "2m", "5h", "3d") for narrow displays
-fn format_relative_time_short(time: SystemTime) -> String {
-    let now = SystemTime::now();
-    let duration = match now.duration_since(time) {
-        Ok(d) => d,
-        Err(_) => return "?".to_string(),
-    };
-
-    let secs = duration.as_secs();
-    let mins = secs / 60;
-    let hours = mins / 60;
-    let days = hours / 24;
-
-    if secs < 60 {
-        "now".to_string()
-    } else if mins < 60 {
-        format!("{}m", mins)
-    } else if hours < 24 {
-        format!("{}h", hours)
-    } else if days < 30 {
-        format!("{}d", days)
-    } else {
-        // Use abbreviated date for older files
-        use std::time::UNIX_EPOCH;
-        let timestamp = time
-            .duration_since(UNIX_EPOCH)
-            .unwrap_or_default()
-            .as_secs();
-        format_date_short_from_timestamp(timestamp)
-    }
-}
-
-/// Format timestamp as "M/D" for narrow displays
-fn format_date_short_from_timestamp(timestamp: u64) -> String {
-    let secs_per_day: u64 = 86400;
-    let days_since_epoch = timestamp / secs_per_day;
-
-    let mut year = 1970u32;
-    let mut remaining_days = days_since_epoch as u32;
-
-    loop {
-        let days_in_year = if is_leap_year(year) { 366 } else { 365 };
-        if remaining_days < days_in_year {
-            break;
-        }
-        remaining_days -= days_in_year;
-        year += 1;
-    }
-
-    let months = [
-        31,
-        if is_leap_year(year) { 29 } else { 28 },
-        31,
-        30,
-        31,
-        30,
-        31,
-        31,
-        30,
-        31,
-        30,
-        31,
-    ];
-
-    let mut month = 1u32;
-    let mut day = remaining_days + 1;
-
-    for (i, &days) in months.iter().enumerate() {
-        if remaining_days < days {
-            month = (i + 1) as u32;
-            day = remaining_days + 1;
-            break;
-        }
-        remaining_days -= days;
-    }
-
-    format!("{}/{}", month, day)
-}
-
-/// Format timestamp as "Mon DD" or "Mon DD YYYY" if not current year
-fn format_date_from_timestamp(timestamp: u64) -> String {
-    // Simple month calculation (approximate, but good enough for display)
-    let secs_per_day: u64 = 86400;
-    let days_since_epoch = timestamp / secs_per_day;
-
-    // Calculate year, month, day (simplified)
-    let mut year = 1970u32;
-    let mut remaining_days = days_since_epoch as u32;
-
-    loop {
-        let days_in_year = if is_leap_year(year) { 366 } else { 365 };
-        if remaining_days < days_in_year {
-            break;
-        }
-        remaining_days -= days_in_year;
-        year += 1;
-    }
-
-    let months = [
-        ("Jan", 31),
-        ("Feb", if is_leap_year(year) { 29 } else { 28 }),
-        ("Mar", 31),
-        ("Apr", 30),
-        ("May", 31),
-        ("Jun", 30),
-        ("Jul", 31),
-        ("Aug", 31),
-        ("Sep", 30),
-        ("Oct", 31),
-        ("Nov", 30),
-        ("Dec", 31),
-    ];
-
-    let mut month_name = "Jan";
-    let mut day = remaining_days + 1;
-
-    for (name, days) in months.iter() {
-        if remaining_days < *days {
-            month_name = name;
-            day = remaining_days + 1;
-            break;
-        }
-        remaining_days -= days;
-    }
-
-    // Get current year for comparison
-    let now_timestamp = SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .unwrap_or_default()
-        .as_secs();
-    let current_year = 1970 + (now_timestamp / (365 * secs_per_day)) as u32;
-
-    if year == current_year {
-        format!("{} {}", month_name, day)
-    } else {
-        format!("{} {} {}", month_name, day, year)
-    }
-}
-
-fn is_leap_year(year: u32) -> bool {
-    (year.is_multiple_of(4) && !year.is_multiple_of(100)) || year.is_multiple_of(400)
-}
-
-/// Render input popup based on current mode
-pub fn render_input_popup(frame: &mut Frame, state: &AppState) {
-    match &state.mode {
-        ViewMode::Input {
-            purpose,
-            buffer,
-            cursor: _,
-        } => {
-            let title = match purpose {
-                InputPurpose::CreateFile => "New File",
-                InputPurpose::CreateDir => "New Directory",
-                InputPurpose::Rename { .. } => "Rename",
-            };
-            draw_input_popup(frame, title, buffer);
-        }
-        ViewMode::Search { query } => {
-            draw_input_popup(frame, "Search", query);
-        }
-        ViewMode::Confirm { action } => {
-            draw_confirm_popup(frame, action);
-        }
-        ViewMode::BookmarkSet => {
-            draw_mini_popup(frame, "Set bookmark (1-9)");
-        }
-        ViewMode::BookmarkJump => {
-            draw_mini_popup(frame, "Jump to bookmark (1-9)");
-        }
-        ViewMode::Filter { query } => {
-            draw_input_popup(frame, "Filter (e.g., *.rs)", query);
-        }
-        _ => {}
-    }
-}
-
-/// Draw a simple input popup
-fn draw_input_popup(frame: &mut Frame, title: &str, content: &str) {
-    let t = theme();
-    let area = centered_rect(60, 3, frame.area());
-
-    let input = Paragraph::new(content)
-        .style(Style::default().fg(t.warning))
-        .block(Block::default().borders(Borders::ALL).title(title));
-
-    frame.render_widget(Clear, area);
-    frame.render_widget(input, area);
-}
-
-/// Draw a small notification popup (for bookmark modes, etc.)
-fn draw_mini_popup(frame: &mut Frame, message: &str) {
-    let t = theme();
-    let width = (message.len() + 4).min(50) as u16;
-    let area = centered_rect(width, 3, frame.area());
-
-    let popup = Paragraph::new(message)
-        .style(Style::default().fg(t.border_active))
-        .block(Block::default().borders(Borders::ALL));
-
-    frame.render_widget(Clear, area);
-    frame.render_widget(popup, area);
-}
-
-/// Draw confirmation popup
-fn draw_confirm_popup(frame: &mut Frame, action: &PendingAction) {
-    match action {
-        PendingAction::Delete { targets } => {
-            draw_delete_confirm_popup(frame, targets);
-        }
-    }
-}
-
-/// Draw delete confirmation popup
-fn draw_delete_confirm_popup(frame: &mut Frame, paths: &[std::path::PathBuf]) {
-    let max_items_to_show = 8;
-    let items_count = paths.len().min(max_items_to_show);
-    let has_more = paths.len() > max_items_to_show;
-    let has_directories = paths.iter().any(|p| p.is_dir());
-
-    let warning_lines = if has_directories { 2 } else { 0 };
-    let more_line = if has_more { 1 } else { 0 };
-    let height = (3 + warning_lines + items_count + more_line + 2) as u16;
-
-    let area = centered_rect(60, height, frame.area());
-
-    let mut content = Vec::new();
-
-    if has_directories {
-        content.push(Line::from(vec![Span::styled(
-            "!! WARNING: FOLDER MOVE !!",
-            Style::default().fg(Color::Red).add_modifier(Modifier::BOLD),
-        )]));
-        content.push(Line::from(vec![Span::styled(
-            "Folders and all contents will be moved to trash",
-            Style::default().fg(Color::Yellow),
-        )]));
-        content.push(Line::from(""));
-    }
-
-    content.push(Line::from(vec![Span::styled(
-        format!("Move {} item(s) to trash:", paths.len()),
-        Style::default().add_modifier(Modifier::BOLD),
-    )]));
-
-    for path in paths.iter().take(max_items_to_show) {
-        let name = path
-            .file_name()
-            .map(|n| n.to_string_lossy().to_string())
-            .unwrap_or_else(|| path.display().to_string());
-
-        let style = if path.is_dir() {
-            Style::default().fg(Color::Red).add_modifier(Modifier::BOLD)
-        } else {
-            Style::default().fg(Color::White)
-        };
-
-        content.push(Line::from(vec![Span::raw("  "), Span::styled(name, style)]));
-    }
-
-    if has_more {
-        content.push(Line::from(vec![Span::styled(
-            format!("  ... and {} more", paths.len() - max_items_to_show),
-            Style::default().fg(Color::DarkGray),
-        )]));
-    }
-
-    content.push(Line::from(""));
-    content.push(Line::from(vec![
-        Span::styled(
-            "y",
-            Style::default()
-                .fg(Color::Green)
-                .add_modifier(Modifier::BOLD),
-        ),
-        Span::raw(" to confirm, "),
-        Span::styled(
-            "n",
-            Style::default().fg(Color::Red).add_modifier(Modifier::BOLD),
-        ),
-        Span::raw(" to cancel"),
-    ]));
-
-    let title = if has_directories {
-        " !! MOVE FOLDERS TO TRASH !! "
-    } else {
-        " Move to Trash "
-    };
-
-    let title_style = if has_directories {
-        Style::default().fg(Color::Red).add_modifier(Modifier::BOLD)
-    } else {
-        Style::default()
-    };
-
-    let popup = Paragraph::new(content).block(
-        Block::default()
-            .borders(Borders::ALL)
-            .border_style(if has_directories {
-                Style::default().fg(Color::Red)
-            } else {
-                Style::default()
-            })
-            .title(Span::styled(title, title_style)),
-    );
-
-    frame.render_widget(Clear, area);
-    frame.render_widget(popup, area);
-}
-
-/// Create a centered rectangle
-fn centered_rect(percent_x: u16, height: u16, area: Rect) -> Rect {
-    let popup_layout = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints([
-            Constraint::Length((area.height.saturating_sub(height)) / 2),
-            Constraint::Length(height),
-            Constraint::Min(0),
-        ])
-        .split(area);
-
-    Layout::default()
-        .direction(Direction::Horizontal)
-        .constraints([
-            Constraint::Percentage((100 - percent_x) / 2),
-            Constraint::Percentage(percent_x),
-            Constraint::Percentage((100 - percent_x) / 2),
-        ])
-        .split(popup_layout[1])[1]
-}
-
-// Help overlay constants
-const HELP_OVERLAY_BG_COLOR: Color = Color::Rgb(30, 30, 40);
-const HELP_OVERLAY_MARGIN: u16 = 1;
-const HELP_MIN_WIDTH: u16 = 24;
-
-#[derive(Debug, Clone, Copy)]
-enum HelpKeyStyle {
-    Solid,
-    Outline,
-    Plain,
-}
-
-fn help_key_style() -> HelpKeyStyle {
-    static STYLE: OnceLock<HelpKeyStyle> = OnceLock::new();
-    *STYLE.get_or_init(|| {
-        match std::env::var("FILEVIEW_HELP_KEY_STYLE")
-            .ok()
-            .map(|s| s.trim().to_ascii_lowercase())
-            .as_deref()
-        {
-            Some("outline") => HelpKeyStyle::Outline,
-            Some("plain") => HelpKeyStyle::Plain,
-            _ => HelpKeyStyle::Solid,
-        }
-    })
-}
-
-/// Create a styled key span with background highlight
-fn help_key(key: &str) -> Span<'_> {
-    let t = theme();
-    let style = match help_key_style() {
-        HelpKeyStyle::Solid => Style::default()
-            .fg(t.help_key_fg)
-            .bg(t.help_key_bg)
-            .add_modifier(Modifier::BOLD),
-        HelpKeyStyle::Outline => Style::default()
-            .fg(t.help_key_bg)
-            .add_modifier(Modifier::BOLD | Modifier::UNDERLINED),
-        HelpKeyStyle::Plain => Style::default().fg(t.info),
-    };
-
-    Span::styled(key, style)
-}
-
-/// Create a description span
-fn help_desc(desc: &str) -> Span<'_> {
-    Span::styled(desc, Style::default().fg(Color::White))
-}
-
-/// Create a section header
-fn help_section(title: &str) -> Line<'_> {
-    Line::from(Span::styled(
-        title,
-        Style::default()
-            .fg(Color::Yellow)
-            .add_modifier(Modifier::BOLD),
-    ))
-}
-
-/// Build help content for narrow width (< 45 chars) - vertical layout
-fn build_narrow_help() -> Vec<Line<'static>> {
-    vec![
-        help_section("Navigation"),
-        Line::from(vec![help_key(" j "), help_key(" ↓ "), help_desc(" Down")]),
-        Line::from(vec![help_key(" k "), help_key(" ↑ "), help_desc(" Up")]),
-        Line::from(vec![
-            help_key(" g "),
-            help_desc(" Top  "),
-            help_key(" G "),
-            help_desc(" Bottom"),
-        ]),
-        Line::from(""),
-        help_section("Tree"),
-        Line::from(vec![help_key(" l "), help_key(" → "), help_desc(" Expand")]),
-        Line::from(vec![
-            help_key(" h "),
-            help_key(" ← "),
-            help_desc(" Collapse"),
-        ]),
-        Line::from(vec![
-            help_key(" L "),
-            help_desc(" All  "),
-            help_key(" H "),
-            help_desc(" All"),
-        ]),
-        Line::from(vec![
-            help_key(" Tab "),
-            help_key(" Enter "),
-            help_desc(" Toggle"),
-        ]),
-        Line::from(""),
-        help_section("Selection"),
-        Line::from(vec![help_key(" Space "), help_desc(" Mark")]),
-        Line::from(vec![help_key(" ^G "), help_desc(" Git changed")]),
-        Line::from(vec![help_key(" ^T "), help_desc(" Test pair")]),
-        Line::from(""),
-        help_section("File"),
-        Line::from(vec![
-            help_key(" a "),
-            help_desc(" File "),
-            help_key(" A "),
-            help_desc(" Dir"),
-        ]),
-        Line::from(vec![help_key(" r "), help_desc(" Rename")]),
-        Line::from(vec![
-            help_key(" y "),
-            help_desc(" Cp "),
-            help_key(" d "),
-            help_desc(" Cut "),
-            help_key(" p "),
-            help_desc(" Paste"),
-        ]),
-        Line::from(vec![help_key(" D "), help_desc(" Delete")]),
-        Line::from(""),
-        help_section("Clipboard"),
-        Line::from(vec![
-            help_key(" c "),
-            help_desc(" Path "),
-            help_key(" C "),
-            help_desc(" Name"),
-        ]),
-        Line::from(vec![help_key(" Y "), help_desc(" Content")]),
-        Line::from(vec![help_key(" ^Y "), help_desc(" Claude fmt")]),
-        Line::from(""),
-        help_section("Search"),
-        Line::from(vec![help_key(" / "), help_desc(" Search")]),
-        Line::from(vec![
-            help_key(" n "),
-            help_desc(" Next "),
-            help_key(" N "),
-            help_desc(" Prev"),
-        ]),
-        Line::from(vec![help_key(" ^P "), help_desc(" Fuzzy")]),
-        Line::from(vec![
-            help_key(" F "),
-            help_desc(" Filter "),
-            help_key(" S "),
-            help_desc(" Sort"),
-        ]),
-        Line::from(""),
-        help_section("Preview"),
-        Line::from(vec![
-            help_key(" o "),
-            help_desc(" Open "),
-            help_key(" P "),
-            help_desc(" Quick"),
-        ]),
-        Line::from(vec![help_key(" b "), help_key(" f "), help_desc(" Scroll")]),
-        Line::from(vec![
-            help_key(" [ "),
-            help_key(" ] "),
-            help_desc(" PDF page"),
-        ]),
-        Line::from(""),
-        help_section("Git"),
-        Line::from(vec![
-            help_key(" s "),
-            help_desc(" Stage "),
-            help_key(" u "),
-            help_desc(" Unstage"),
-        ]),
-        Line::from(""),
-        help_section("Bookmarks"),
-        Line::from(vec![help_key(" m "), help_desc("+1-9 Set")]),
-        Line::from(vec![help_key(" ' "), help_desc("+1-9 Jump")]),
-        Line::from(""),
-        help_section("Tabs"),
-        Line::from(vec![
-            help_key(" ^T "),
-            help_desc(" New "),
-            help_key(" ^W "),
-            help_desc(" Close"),
-        ]),
-        Line::from(vec![
-            help_key(" A-t "),
-            help_desc(" Next "),
-            help_key(" A-T "),
-            help_desc(" Prev"),
-        ]),
-        Line::from(""),
-        help_section("Other"),
-        Line::from(vec![
-            help_key(" . "),
-            help_desc(" Hidden "),
-            help_key(" F5 "),
-            help_desc(" Refresh"),
-        ]),
-        Line::from(vec![
-            help_key(" q "),
-            help_desc(" Quit "),
-            help_key(" Q "),
-            help_desc(" Quit+cd"),
-        ]),
-        Line::from(vec![
-            help_key(" ? "),
-            help_desc(" Help "),
-            help_key(" Esc "),
-            help_desc(" Cancel"),
-        ]),
-        Line::from(""),
-        Line::from(Span::styled(
-            "? or Esc to close",
-            Style::default().fg(Color::DarkGray),
-        )),
-    ]
-}
-
-/// Build help content for wide width (>= 45 chars) - horizontal layout
-fn build_wide_help() -> Vec<Line<'static>> {
-    vec![
-        help_section("Navigation"),
-        Line::from(vec![
-            help_key(" j "),
-            help_desc("/"),
-            help_key(" ↓ "),
-            help_desc(" Down   "),
-            help_key(" k "),
-            help_desc("/"),
-            help_key(" ↑ "),
-            help_desc(" Up   "),
-            help_key(" g "),
-            help_desc(" Top   "),
-            help_key(" G "),
-            help_desc(" Bottom"),
-        ]),
-        Line::from(""),
-        help_section("Tree"),
-        Line::from(vec![
-            help_key(" l "),
-            help_desc("/"),
-            help_key(" → "),
-            help_desc(" Expand   "),
-            help_key(" h "),
-            help_desc("/"),
-            help_key(" ← "),
-            help_desc(" Collapse   "),
-            help_key(" Tab "),
-            help_desc(" Toggle"),
-        ]),
-        Line::from(vec![
-            help_key(" L "),
-            help_desc(" Expand all   "),
-            help_key(" H "),
-            help_desc(" Collapse all   "),
-            help_key(" Enter "),
-            help_desc(" Toggle/Pick"),
-        ]),
-        Line::from(""),
-        help_section("Selection"),
-        Line::from(vec![
-            help_key(" Space "),
-            help_desc(" Mark   "),
-            help_key(" Ctrl+G "),
-            help_desc(" Git changed   "),
-            help_key(" Ctrl+T "),
-            help_desc(" Test pair"),
-        ]),
-        Line::from(""),
-        help_section("File Operations"),
-        Line::from(vec![
-            help_key(" a "),
-            help_desc(" New file   "),
-            help_key(" A "),
-            help_desc(" New dir   "),
-            help_key(" r "),
-            help_desc(" Rename   "),
-            help_key(" R "),
-            help_desc(" Bulk rename"),
-        ]),
-        Line::from(vec![
-            help_key(" y "),
-            help_desc(" Copy   "),
-            help_key(" d "),
-            help_desc(" Cut   "),
-            help_key(" p "),
-            help_desc(" Paste   "),
-            help_key(" D "),
-            help_desc("/"),
-            help_key(" Del "),
-            help_desc(" Delete"),
-        ]),
-        Line::from(""),
-        help_section("Clipboard"),
-        Line::from(vec![
-            help_key(" c "),
-            help_desc(" Path   "),
-            help_key(" C "),
-            help_desc(" Filename   "),
-            help_key(" Y "),
-            help_desc(" Content   "),
-            help_key(" Ctrl+Y "),
-            help_desc(" Claude format"),
-        ]),
-        Line::from(""),
-        help_section("Search & Filter"),
-        Line::from(vec![
-            help_key(" / "),
-            help_desc(" Search   "),
-            help_key(" n "),
-            help_desc(" Next   "),
-            help_key(" N "),
-            help_desc(" Prev   "),
-            help_key(" Ctrl+P "),
-            help_desc(" Fuzzy finder"),
-        ]),
-        Line::from(vec![
-            help_key(" F "),
-            help_desc(" Filter   "),
-            help_key(" S "),
-            help_desc(" Sort mode"),
-        ]),
-        Line::from(""),
-        help_section("Preview"),
-        Line::from(vec![
-            help_key(" o "),
-            help_desc(" Full preview   "),
-            help_key(" P "),
-            help_desc(" Quick preview"),
-        ]),
-        Line::from(vec![
-            help_key(" b "),
-            help_desc("/"),
-            help_key(" PgUp "),
-            help_desc(" Up   "),
-            help_key(" f "),
-            help_desc("/"),
-            help_key(" PgDn "),
-            help_desc(" Down   "),
-            help_key(" [ "),
-            help_key(" ] "),
-            help_desc(" PDF pages"),
-        ]),
-        Line::from(""),
-        help_section("Git"),
-        Line::from(vec![
-            help_key(" s "),
-            help_desc(" Stage   "),
-            help_key(" u "),
-            help_desc(" Unstage"),
-        ]),
-        Line::from(""),
-        help_section("Bookmarks"),
-        Line::from(vec![
-            help_key(" m "),
-            help_desc("+"),
-            help_key(" 1-9 "),
-            help_desc(" Set   "),
-            help_key(" ' "),
-            help_desc("+"),
-            help_key(" 1-9 "),
-            help_desc(" Jump"),
-        ]),
-        Line::from(""),
-        help_section("Tabs"),
-        Line::from(vec![
-            help_key(" Ctrl+T "),
-            help_desc(" New   "),
-            help_key(" Ctrl+W "),
-            help_desc(" Close   "),
-            help_key(" Alt+t "),
-            help_desc(" Next   "),
-            help_key(" Alt+T "),
-            help_desc(" Prev"),
-        ]),
-        Line::from(""),
-        help_section("Other"),
-        Line::from(vec![
-            help_key(" . "),
-            help_desc(" Hidden   "),
-            help_key(" F5 "),
-            help_desc(" Refresh   "),
-            help_key(" ? "),
-            help_desc(" Help   "),
-            help_key(" q "),
-            help_desc(" Quit   "),
-            help_key(" Q "),
-            help_desc(" Quit+cd"),
-        ]),
-        Line::from(""),
-        Line::from(Span::styled(
-            "  Press ? or Esc to close",
-            Style::default().fg(Color::DarkGray),
-        )),
-    ]
-}
-
-/// Render help popup overlay with all keybindings (responsive)
-pub fn render_help_popup(frame: &mut Frame, state: &AppState) {
-    if !matches!(state.mode, ViewMode::Help) {
-        return;
-    }
-
-    let area = frame.area();
-
-    // Calculate overlay size - fit to screen with margins
-    let overlay_width = area
-        .width
-        .saturating_sub(HELP_OVERLAY_MARGIN * 2)
-        .max(HELP_MIN_WIDTH);
-    let overlay_height = area.height.saturating_sub(HELP_OVERLAY_MARGIN * 2).max(10);
-    let overlay_x = (area.width.saturating_sub(overlay_width)) / 2;
-    let overlay_y = (area.height.saturating_sub(overlay_height)) / 2;
-
-    let overlay_area = Rect::new(overlay_x, overlay_y, overlay_width, overlay_height);
-
-    // Choose layout based on available width
-    let inner_width = overlay_width.saturating_sub(2);
-    let content = if inner_width < 45 {
-        build_narrow_help()
-    } else {
-        build_wide_help()
-    };
-
-    // Clear background and render
-    frame.render_widget(Clear, overlay_area);
-
-    let paragraph = Paragraph::new(content).block(
-        Block::default()
-            .title(" Keybindings ")
-            .borders(Borders::ALL)
-            .border_style(Style::default().fg(Color::Cyan))
-            .style(Style::default().bg(HELP_OVERLAY_BG_COLOR)),
-    );
-
-    frame.render_widget(paragraph, overlay_area);
-}
-
-/// Render the TODO/FIXME aggregator popup.
-pub fn render_todos_popup(frame: &mut Frame, state: &AppState) {
-    use crate::integrate::TodoTag;
-
-    let selected = match state.mode {
-        ViewMode::Todos { selected } => selected,
-        _ => return,
-    };
-
-    let area = frame.area();
-    let overlay_width = area.width.saturating_sub(8).max(60);
-    let overlay_height = area.height.saturating_sub(4).max(10);
-    let overlay_x = (area.width.saturating_sub(overlay_width)) / 2;
-    let overlay_y = (area.height.saturating_sub(overlay_height)) / 2;
-    let overlay_area = Rect::new(overlay_x, overlay_y, overlay_width, overlay_height);
-
-    frame.render_widget(Clear, overlay_area);
-
-    let t = theme();
-    let inner_width = overlay_width.saturating_sub(2) as usize;
-    let visible_rows = overlay_height.saturating_sub(2) as usize;
-
-    // Center the selected row in the visible window.
-    let total = state.todo_items.len();
-    let start = if total <= visible_rows {
-        0
-    } else {
-        let half = visible_rows / 2;
-        if selected < half {
-            0
-        } else if selected + (visible_rows - half) >= total {
-            total.saturating_sub(visible_rows)
-        } else {
-            selected.saturating_sub(half)
-        }
-    };
-
-    let mut lines: Vec<Line<'static>> = Vec::with_capacity(visible_rows);
-    if total == 0 {
-        let msg = if state.todo_partial {
-            "No TODOs in scanned region (scan was partial)"
-        } else {
-            "No TODOs found"
-        };
-        lines.push(Line::from(Span::styled(
-            msg.to_string(),
-            Style::default().fg(t.git_ignored),
-        )));
-    } else {
-        for (i, item) in state
-            .todo_items
-            .iter()
-            .enumerate()
-            .skip(start)
-            .take(visible_rows)
-        {
-            let tag_color = match item.tag {
-                TodoTag::Todo => t.info,
-                TodoTag::Fixme => t.warning,
-                TodoTag::Xxx => t.warning,
-                TodoTag::Hack => t.git_conflict,
-                TodoTag::Bug => t.git_conflict,
-                TodoTag::Note => t.git_staged,
-            };
-            let rel = item
-                .path
-                .strip_prefix(&state.root)
-                .unwrap_or(&item.path)
-                .display()
-                .to_string();
-            let prefix = if i == selected { "▶ " } else { "  " };
-            let location = format!("{}:{}", rel, item.line);
-            let msg = if item.message.is_empty() {
-                String::new()
-            } else {
-                format!("  {}", item.message)
-            };
-
-            let line_style = if i == selected {
-                Style::default()
-                    .bg(t.selection)
-                    .add_modifier(Modifier::BOLD)
-            } else {
-                Style::default()
-            };
-
-            // Truncate the rendered text so it never overflows the popup width.
-            let rendered = format!("{}{:6} {}{}", prefix, item.tag.as_str(), location, msg);
-            let truncated = if rendered.chars().count() > inner_width {
-                let mut out: String = rendered
-                    .chars()
-                    .take(inner_width.saturating_sub(1))
-                    .collect();
-                out.push('…');
-                out
-            } else {
-                rendered
-            };
-
-            // Build line with the tag colored even when the row is selected.
-            let tag_part = format!("{}{:6} ", prefix, item.tag.as_str());
-            let rest_start = tag_part.chars().count().min(truncated.chars().count());
-            let rest: String = truncated.chars().skip(rest_start).collect();
-
-            lines.push(Line::from(vec![
-                Span::styled(tag_part, line_style.fg(tag_color)),
-                Span::styled(rest, line_style),
-            ]));
-        }
-    }
-
-    let title = if state.todo_partial {
-        format!(
-            " TODOs · {} (partial) · Esc to close ",
-            state.todo_items.len()
-        )
-    } else {
-        format!(" TODOs · {} · Esc to close ", state.todo_items.len())
-    };
-
-    let paragraph = Paragraph::new(lines).block(
-        Block::default()
-            .title(title)
-            .borders(Borders::ALL)
-            .border_style(Style::default().fg(t.border_active)),
-    );
-
-    frame.render_widget(paragraph, overlay_area);
-}
